refactor: Phase 2 - Move instruction disassembly to arch/aarch64
Move ARM64-specific instruction disassembly code from call_graph.rs
to arch/aarch64 module:

- Constants: INSN_SIZE, MIN_CHUNK_SIZE, BL_MASK, BL_OPCODE, B_MASK, B_OPCODE
- Struct: InsnData
- Functions:
  - decode_branch_target()
  - scan_branch_instructions()
  - parallel_disassemble_arm64() → parallel_disassemble()
  - sequential_disassemble_arm64() → sequential_disassemble()
- Tests: All tests for above functions moved to arch/aarch64 module

Updated call_graph.rs to use arch::parallel_disassemble() and
arch::sequential_disassemble().

All 406 unit tests pass.

Related: #238
