+++ jonesy/src/call_graph.rs
@@ -1,11 +1,10 @@
+use crate::arch;
 use crate::binary_format::BinaryRef;
 use crate::crate_line_table::CrateLineTable;
 use crate::full_line_table::FullLineTable;
 use crate::function_index::{FunctionIndex, get_functions_from_dwarf, load_dwarf_sections};
 use crate::project_context::ProjectContext;
 use crate::sym::SymbolIndex;
-use capstone::arch::BuildsCapstone;
-use capstone::{Capstone, arch};
 use dashmap::DashMap;
 use goblin::elf::Elf;
 use goblin::mach::MachO;
@@ -34,26 +33,6 @@ pub struct CallerInfo<'a> {
     /// Column number of the calling instruction
     pub column: Option<u32>,
 }
-
-/// Extracted instruction data for parallel processing (avoids Insn lifetime issues)
-struct InsnData {
-    address: u64,
-    call_target: Option<u64>,
-}
-
-/// ARM64 instruction size in bytes (fixed-size ISA)
-const ARM64_INSN_SIZE: usize = 4;
-
-/// Minimum chunk size for parallel disassembly (avoid overhead for small sections)
-const MIN_CHUNK_SIZE: usize = 64 * 1024; // 64KB
-
-/// ARM64 BL instruction mask: bits [31:26] must be 100101
-const BL_MASK: u32 = 0xFC000000;
-const BL_OPCODE: u32 = 0x94000000;
-
-/// ARM64 B instruction mask: bits [31:26] must be 000101
-const B_MASK: u32 = 0xFC000000;
-const B_OPCODE: u32 = 0x14000000;
 
 /// Build a mapping from PLT stub addresses to actual function addresses using ELF relocations.
 /// Returns a HashMap mapping PLT address -> target function address.
@@ -152,124 +131,6 @@ fn resolve_plt_stub(target_addr: u64, plt_map: &HashMap<u64, u64>) -> u64 {
     plt_map.get(&target_addr).copied().unwrap_or(target_addr)
 }
 
-/// Decode ARM64 BL/B instruction target address from raw bytes.
-/// BL encoding: 100101 imm26
-/// B encoding: 000101 imm26
-/// Target = PC + sign_extend(imm26) * 4
-fn decode_branch_target(insn_bytes: u32, pc: u64) -> u64 {
-    // Extract 26-bit immediate
-    let imm26 = insn_bytes & 0x03FFFFFF;
-    // Sign-extend to 32 bits and multiply by 4 (shift left 2)
-    let offset = ((imm26 as i32) << 6) >> 4;
-    // Add to PC
-    (pc as i64 + offset as i64) as u64
-}
-
-/// Scan for ARM64 BL and B instructions in parallel by dividing into chunks.
-/// BL = branch with link (function calls)
-/// B = unconditional branch (tail calls to other functions)
-/// Directly scans raw bytes for patterns - no disassembly needed.
-/// This is much faster than using Capstone for full disassembly.
-fn parallel_disassemble_arm64(text_data: &[u8], text_addr: u64) -> Vec<InsnData> {
-    let num_threads = rayon::current_num_threads();
-
-    // Calculate chunk size, ensuring alignment to instruction boundary
-    let ideal_chunk_size = text_data.len() / num_threads;
-    let chunk_size = if ideal_chunk_size < MIN_CHUNK_SIZE {
-        // Data too small to benefit from parallelization
-        text_data.len()
-    } else {
-        // Align to 4-byte instruction boundary
-        (ideal_chunk_size / ARM64_INSN_SIZE) * ARM64_INSN_SIZE
-    };
-
-    if chunk_size >= text_data.len() {
-        // Single chunk - use sequential scanning
-        return scan_branch_instructions(text_data, text_addr);
-    }
-
-    // Create chunks with their base addresses
-    let chunks: Vec<(usize, &[u8], u64)> = text_data
-        .chunks(chunk_size)
-        .enumerate()
-        .map(|(i, chunk)| {
-            let chunk_addr = text_addr + (i * chunk_size) as u64;
-            (i, chunk, chunk_addr)
-        })
-        .collect();
-
-    // Scan chunks in parallel for BL and B instructions
-    let results: Vec<Vec<InsnData>> = chunks
-        .par_iter()
-        .map(|(_i, chunk, chunk_addr)| scan_branch_instructions(chunk, *chunk_addr))
-        .collect();
-
-    // Flatten results from all chunks
-    results.into_iter().flatten().collect()
-}
-
-/// Scan a chunk of ARM64 code for BL and B instructions.
-/// BL = branch with link (function calls)
-/// B = unconditional branch (tail calls to other functions)
-/// Directly checks raw bytes against opcode patterns.
-fn scan_branch_instructions(data: &[u8], base_addr: u64) -> Vec<InsnData> {
-    data.chunks_exact(ARM64_INSN_SIZE)
-        .enumerate()
-        .filter_map(|(i, bytes)| {
-            let insn = u32::from_le_bytes([bytes[0], bytes[1], bytes[2], bytes[3]]);
-            let is_bl = (insn & BL_MASK) == BL_OPCODE;
-            let is_b = (insn & B_MASK) == B_OPCODE;
-            if is_bl || is_b {
-                let pc = base_addr + (i * ARM64_INSN_SIZE) as u64;
-                let target = decode_branch_target(insn, pc);
-                Some(InsnData {
-                    address: pc,
-                    call_target: Some(target),
-                })
-            } else {
-                None
-            }
-        })
-        .collect()
-}
-
-/// Sequential disassembly using Capstone - kept for non-ARM64 platforms.
-#[allow(dead_code)]
-fn sequential_disassemble_arm64(text_data: &[u8], text_addr: u64) -> Vec<InsnData> {
-    let Ok(cs) = Capstone::new()
-        .arm64()
-        .mode(arch::arm64::ArchMode::Arm)
-        .build()
-    else {
-        eprintln!("Warning: failed to initialize Capstone disassembler");
-        return Vec::new();
-    };
-
-    let Ok(instructions) = cs.disasm_all(text_data, text_addr) else {
-        eprintln!("Warning: disassembly failed for text section at {text_addr:#x}");
-        return Vec::new();
-    };
-
-    instructions
-        .iter()
-        .filter_map(|insn| {
-            // Match both BL (branch with link) and B (branch) for tail call detection
-            let mnemonic = insn.mnemonic();
-            if mnemonic == Some("bl") || mnemonic == Some("b") {
-                let operand = insn.op_str()?;
-                let addr_str = operand.trim_start_matches("#0x");
-                let call_target = u64::from_str_radix(addr_str, 16).ok();
-                Some(InsnData {
-                    address: insn.address(),
-                    call_target,
-                })
-            } else {
-                None
-            }
-        })
-        .collect()
-}
-
 /// Get a named section's address and data from a binary.
 /// This is a standalone helper that wraps BinaryRef::find_section for callers
 /// that only have a raw MachO reference. For new code, use BinaryRef::find_section directly.
@@ -316,10 +177,10 @@ impl<'a> CallGraph<'a> {
 
         // Parallel disassembly - divides text section into chunks (ARM64 only)
         #[cfg(target_arch = "aarch64")]
-        let insn_data = parallel_disassemble_arm64(text_data, text_addr);
+        let insn_data = arch::parallel_disassemble(text_data, text_addr);
 
         #[cfg(not(target_arch = "aarch64"))]
-        let insn_data = sequential_disassemble_arm64(text_data, text_addr);
+        let insn_data = arch::sequential_disassemble(text_data, text_addr);
 
         // Process bl instructions in parallel (the expensive part is function lookup)
         let edges: DashMap<u64, Vec<CallerInfo<'a>>> = DashMap::new();
@@ -413,10 +274,10 @@ impl<'a> CallGraph<'a> {
         // Parallel disassembly - divides text section into chunks (ARM64 only)
         let step = Instant::now();
         #[cfg(target_arch = "aarch64")]
-        let insn_data = parallel_disassemble_arm64(text_data, text_addr);
+        let insn_data = arch::parallel_disassemble(text_data, text_addr);
 
         #[cfg(not(target_arch = "aarch64"))]
-        let insn_data = sequential_disassemble_arm64(text_data, text_addr);
+        let insn_data = arch::sequential_disassemble(text_data, text_addr);
         if show_timings {
             // insn_data contains only BL/B branch instructions (not all instructions)
             eprintln!(
@@ -494,7 +355,7 @@ impl<'a> CallGraph<'a> {
 /// Falls back to symbol table lookup if DWARF doesn't contain the function.
 /// DWARF names use Cow::Owned, symbol fallback uses Cow::Borrowed from SymbolIndex.
 fn process_instruction_data_with_crate_table<'a>(
-    data: &InsnData,
+    data: &arch::InsnData,
     function_index: &FunctionIndex,
     full_line_table: &FullLineTable,
     crate_line_table: &CrateLineTable,
@@ -632,37 +493,6 @@ fn process_instruction_data_with_crate_table<'a>(
 mod tests {
     use super::*;
 
-    // Tests for decode_branch_target (ARM64)
-    #[test]
-    fn test_decode_branch_target_forward() {
-        // BL instruction: offset = +4 (next instruction)
-        // imm26 = 1, pc = 0x1000
-        let insn = 0x94000001_u32; // BL +4
-        let target = decode_branch_target(insn, 0x1000);
-        assert_eq!(target, 0x1004);
-    }
-
-    #[test]
-    fn test_decode_branch_target_backward() {
-        // BL instruction with negative offset (high bit set in imm26)
-        // This represents a backward branch
-        let pc = 0x2000_u64;
-        // imm26 = 0x3FFFFFF (-1 in 26-bit signed) => offset = -4
-        let insn = 0x97FFFFFF_u32; // BL -4
-        let target = decode_branch_target(insn, pc);
-        assert_eq!(target, pc.wrapping_sub(4));
-    }
-
-    #[test]
-    fn test_decode_branch_target_zero_offset() {
-        // BL with offset 0 (branches to itself)
-        let insn = 0x94000000_u32;
-        let target = decode_branch_target(insn, 0x1000);
-        assert_eq!(target, 0x1000);
-    }
-
-    // Additional test for NEW logic in simplify_heuristics branch
-
     #[test]
     fn test_get_section_by_name_nonexistent() {
         // Create a minimal test binary to parse
@@ -703,103 +533,6 @@ mod tests {
         let graph: CallGraph<'_> = CallGraph::empty();
         let callers = graph.get_callers(0xDEADBEEF);
         assert!(callers.is_empty());
-    }
-
-    // Tests for scan_branch_instructions
-    #[test]
-    fn test_scan_branch_instructions_bl() {
-        // BL +4 instruction at address 0x1000
-        let bl_insn: [u8; 4] = 0x94000001_u32.to_le_bytes();
-        let results = scan_branch_instructions(&bl_insn, 0x1000);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].address, 0x1000);
-        assert_eq!(results[0].call_target, Some(0x1004));
-    }
-
-    #[test]
-    fn test_scan_branch_instructions_b() {
-        // B +8 instruction (unconditional branch / tail call)
-        let b_insn: [u8; 4] = 0x14000002_u32.to_le_bytes();
-        let results = scan_branch_instructions(&b_insn, 0x2000);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].address, 0x2000);
-        assert_eq!(results[0].call_target, Some(0x2008));
-    }
-
-    #[test]
-    fn test_scan_branch_instructions_non_branch() {
-        // ADD instruction (not a branch)
-        let add_insn: [u8; 4] = 0x91000000_u32.to_le_bytes();
-        let results = scan_branch_instructions(&add_insn, 0x1000);
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn test_scan_branch_instructions_multiple() {
-        // Two BL instructions with a non-branch between them
-        let mut code = Vec::new();
-        code.extend_from_slice(&0x94000003_u32.to_le_bytes()); // BL +12
-        code.extend_from_slice(&0x91000000_u32.to_le_bytes()); // ADD (not branch)
-        code.extend_from_slice(&0x94000001_u32.to_le_bytes()); // BL +4
-
-        let results = scan_branch_instructions(&code, 0x1000);
-        assert_eq!(results.len(), 2);
-        assert_eq!(results[0].address, 0x1000);
-        assert_eq!(results[0].call_target, Some(0x100C));
-        assert_eq!(results[1].address, 0x1008);
-        assert_eq!(results[1].call_target, Some(0x100C));
-    }
-
-    #[test]
-    fn test_scan_branch_instructions_empty_input() {
-        let results = scan_branch_instructions(&[], 0x1000);
-        assert!(results.is_empty());
-    }
-
-    // Tests for parallel_disassemble_arm64
-    #[test]
-    fn test_parallel_disassemble_small_input() {
-        // Small input falls through to sequential scan_branch_instructions
-        let mut code = Vec::new();
-        code.extend_from_slice(&0x94000002_u32.to_le_bytes()); // BL +8
-        code.extend_from_slice(&0x91000000_u32.to_le_bytes()); // ADD (not branch)
-        code.extend_from_slice(&0x14000001_u32.to_le_bytes()); // B +4
-
-        let results = parallel_disassemble_arm64(&code, 0x1000);
-        assert_eq!(results.len(), 2);
-        assert_eq!(results[0].address, 0x1000);
-        assert_eq!(results[1].address, 0x1008);
-    }
-
-    #[test]
-    fn test_parallel_disassemble_empty() {
-        let results = parallel_disassemble_arm64(&[], 0x1000);
-        assert!(results.is_empty());
-    }
-
-    // Tests for sequential_disassemble_arm64
-    #[test]
-    fn test_sequential_disassemble_bl_instruction() {
-        // BL +8 at address 0x1000
-        let code: Vec<u8> = 0x94000002_u32.to_le_bytes().to_vec();
-        let results = sequential_disassemble_arm64(&code, 0x1000);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].address, 0x1000);
-        assert!(results[0].call_target.is_some());
-    }
-
-    #[test]
-    fn test_sequential_disassemble_non_branch() {
-        // MOV x0, x1 — not a branch
-        let code: Vec<u8> = 0xAA0103E0_u32.to_le_bytes().to_vec();
-        let results = sequential_disassemble_arm64(&code, 0x1000);
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn test_sequential_disassemble_empty() {
-        let results = sequential_disassemble_arm64(&[], 0x1000);
-        assert!(results.is_empty());
     }
 
     // Tests for CallGraph::build using real binary
